Rename grovers-algo to grovers
