+++ package.json
@@ -74,7 +74,7 @@
       "ibm-quantum-system": "quantum/nodes/ibm-quantum-system/ibm-quantum-system.js",
       "circuit-diagram": "quantum/nodes/circuit-diagram/circuit-diagram.js",
       "script": "quantum/nodes/script/script.js",
-      "grovers":  "quantum/nodes/grovers-algo/grovers.js",
+      "grovers":  "quantum/nodes/grovers/grovers.js",
       "bloch-sphere": "quantum/nodes/bloch-sphere/bloch-sphere.js"
     }
   }
