+++ quantum/nodes/grovers/locales/en-US/grovers.html
@@ -1,4 +1,4 @@
-<script type="text/html" data-help-name="grovers-algo"> <!-- Change name -->
+<script type="text/html" data-help-name="grovers">
   <p>
     This node applies Grover's algorithm to find specified bit string value in the search space
   </p>
